% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-guided_group_lasso.R
\name{BuildTreeHC}
\alias{BuildTreeHC}
\title{Construct tree for \code{\link{TreeGuidedGroupLasso}} using hierarchical
clustering.}
\usage{
BuildTreeHC(Y, threshold = 0.9, linkage = "complete")
}
\arguments{
\item{Y}{N by K matrix for every task.}

\item{threshold}{Threshold used for cutting nodes with high dissimilarity.}

\item{linkage}{Linkage type used for hierarchical clustering.}
}
\value{
List containing \item{groups}{V by K matrix determining group
  membership: Task k in group v iff groups[v,k] == 1.} \item{weights}{V
  dimensional weight vector.}
}
\description{
Compute tree and node weights from the results of applying hierarchical
clustering to the (response) matrix Y.
}
\seealso{
\code{\link{TreeGuidedGroupLasso}},
  \code{\link{ComputeGroupWeights}}.
}
