% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-guided_group_lasso.R
\name{ComputeGroupWeights}
\alias{ComputeGroupWeights}
\title{Compute node weights using scheme from Kim et al [2010].}
\usage{
ComputeGroupWeights(groups, s)
}
\arguments{
\item{groups}{V by K matrix determining group membership: Task k in group v iff groups[v,k] == 1.}

\item{s}{V dimensional vector with weights. Entries for singleton groups will be ignored.}
}
\value{
Vector containing node weights.
}
\description{
Compute node weights using scheme from Kim et al [2010].
}
\seealso{
\code{\link{TreeGuidedGroupLasso}},
  \code{\link{BuildTreeHC}}.
}
