% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tggl_mixture.R
\name{ComputeLogLikelihood}
\alias{ComputeLogLikelihood}
\title{Compute the log-likelihood for a given mixture model.}
\usage{
ComputeLogLikelihood(models, prior, sigmas, X = NULL,
  task.specific.features = list(), Y)
}
\arguments{
\item{models}{List of TGGL models.}

\item{prior}{Vector of prior probabilities for each component.}

\item{sigmas}{M by K Matrix of standard deviations.}

\item{X}{N by J1 matrix of features common to all tasks.}

\item{task.specific.features}{List of features which are specific to each
task. Each entry contains an N by J2 matrix for one particular task (where
columns are features). List has to be ordered according to the columns of
Y.}

\item{Y}{N by K output matrix for every task.}
}
\value{
Log-likelihood.
}
\description{
Compute the log-likelihood for a given mixture model.
}
