% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree-guided_group_lasso.R
\name{ComputeWeightsFromTree}
\alias{ComputeWeightsFromTree}
\title{Compute weights for \code{\link{TreeGuidedGroupLasso}} from prior grouping.}
\usage{
ComputeWeightsFromTree(Y, groups, linkage = "complete",
  distance = function(a, b) {     1 - cor(a, b) })
}
\arguments{
\item{Y}{N by K matrix for every task.}

\item{groups}{V by K matrix determining group membership: Task k in group v
iff groups[v,k] == 1.}

\item{linkage}{"complete" or "average". Linkage type used (distance measure
between clusters).}

\item{distance}{Function for computing pairwise distances between columns of
Y. Default is 1-cor(a,b).}
}
\value{
Vector containing node weights.
}
\description{
Generalize hierarchical clustering distances to compute node weights from the
(response) matrix Y for a given grouping. Weights will be normalized using
ComputeGroupWeights.
}
\seealso{
\code{\link{TreeGuidedGroupLasso}},
  \code{\link{ComputeGroupWeights}}.
}
