% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{EvaluateClusteredLinearMTModel}
\alias{EvaluateClusteredLinearMTModel}
\title{Evaluate clustered LinearMTL model.}
\usage{
EvaluateClusteredLinearMTModel(X = NULL, task.specific.features = list(), Y,
  B.list, train.idx.by.cluster, test.idx.by.cluster)
}
\arguments{
\item{X}{Column centered N by J input matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{Y}{Column centered N by K output matrix for every task.}

\item{B.list}{List of J by K matrices of regression coefficients, J = J1 + J2.}

\item{train.idx.by.cluster}{List of training indices per cluster.}

\item{test.idx.by.cluster}{List of test indices per cluster.}
}
\value{
Matrix containing training and test correlation and MSE.
}
\description{
Compute training and test errors for the given indices.
}
