% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{EvaluateClusteredTreeGuidedGroupLasso}
\alias{EvaluateClusteredTreeGuidedGroupLasso}
\title{Evaluate clustered LinearMTL model.}
\usage{
EvaluateClusteredTreeGuidedGroupLasso(X = NULL,
  task.specific.features = list(), Y, B.list, train.idx.by.cluster,
  test.idx.by.cluster, out.dir = NULL)
}
\arguments{
\item{X}{Column centered N by J input matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{B.list}{List of J by K matrices of regression coefficients, J = J1 + J2.}

\item{train.idx.by.cluster}{List of training indices per cluster.}

\item{test.idx.by.cluster}{List of test indices per cluster.}

\item{out.dir}{Output directory.}
}
\description{
Compute training and test errors for the given indices.
}
