% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{EvaluateLinearMTModel}
\alias{EvaluateLinearMTModel}
\title{Evaluate LinearMTL model.}
\usage{
EvaluateLinearMTModel(X = NULL, task.specific.features = list(), Y,
  LMTL.model, train.idx, test.idx, task.grouping = NULL,
  feature.names = NULL, task.names = NULL, sd.threshold = 1.5, out.dir)
}
\arguments{
\item{X}{Column centered N by J input matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{Y}{Column centered N by K output matrix for every task.}

\item{LMTL.model}{Linear multi-task learning model (list containing B and
intercept).}

\item{train.idx}{Indices for the training set.}

\item{test.idx}{Indices for the test set.}

\item{task.grouping}{String vector of length K with group names for each
task.}

\item{feature.names}{Feature names.}

\item{task.names}{Task names.}

\item{sd.threshold}{All features with error changes above sd.threshold times
the standard deviation will be taken as important.}

\item{out.dir}{Output directory for results and plots.}
}
\description{
Compute training and test errors for the given indices as well as error
changes for excluding features. If task.grouping is supplied, identify top
regulators per group and plot clustering of the regression coefficients along
with the most important features. Save results to directory out.dir.
}
