% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_variable_regression.R
\name{LatentVariableRegression}
\alias{LatentVariableRegression}
\title{Fit a latent variable linar multi-task model (Zhang et al. 2008).}
\usage{
LatentVariableRegression(X = NULL, task.specific.features = list(), Y, S,
  max.iter = 10000, epsilon = 1e-05, XTX = NULL, XTY = NULL,
  verbose = 1)
}
\arguments{
\item{X}{N by J1 matrix of features common to all tasks.}

\item{task.specific.features}{List of features which are specific to each
task. Each entry contains an N by J2 matrix for one particular task (where
columns are features). List has to be ordered according to the columns of
Y.}

\item{Y}{N by K output matrix for every task.}

\item{S}{H by K loading matrix.}

\item{max.iter}{(Optional) Maximum number of iterations.}

\item{epsilon}{(Optional) Desired accuracy. If error change drops below
epsilon, the algorithm terminates.}

\item{XTX}{(Optional) Precomputed matrices t(X)*X as for example produced by
PrepareMatrices.}

\item{XTY}{(Optional) Precomputed matrices t(X)*Y as for example produced by
PrepareMatrices}

\item{verbose}{(Optional) Integer in {0,1,2}. verbose = 0: No output. verbose
= 1: Print summary at the end of the optimization. verbose = 2: Print
progress during optimization.}
}
\value{
List containing
  \item{Gamma}{Estimated mixing matrix.}
  \item{sigma2}{Estimated sigma^2.}
  \item{Psi}{Estimated Psi.}
  \item{S}{Loading matrix used.}
  \item{B}{MAP estimate of the regression coefficients.}
  \item{early.termination}{Boolean indicating whether the algorithm exceeded
  max.iter iterations.}
}
\description{
A Bayesian linear multi-task model where the regression matrix is assumed to
be composed of latent factors.
}
\seealso{
\code{\link{RunGroupCrossvalidation}}
}
