% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_task_utilities.R
\name{MTPredict}
\alias{MTPredict}
\title{Compute predictions for linear multi-task model.}
\usage{
MTPredict(LMTL.model, X = NULL, task.specific.features = list())
}
\arguments{
\item{LMTL.model}{Linear multi-task learning model (list containing B and
intercept).}

\item{X}{N by J1 matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}
}
\value{
N by J predictions matrix.
}
\description{
Compute linear predictions from predictor variables and regression
coefficients.
}
