% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{PlotCustomHeatmap}
\alias{PlotCustomHeatmap}
\title{Plot customized heatmap.}
\usage{
PlotCustomHeatmap(matrix, scale = "none", trace = "none", task.grouping,
  dendro = "col", density.info = "none", ...)
}
\arguments{
\item{matrix}{Matrix to display.}

\item{scale}{See heatmap.2.}

\item{trace}{See heatmap.2}

\item{task.grouping}{Vector of strings with groups for each task.}

\item{dendro}{Dendrogram obtained from clustering the columns of the input matrix.}

\item{density.info}{See heatmap.2.}

\item{...}{Other arguments passed to heatmap.2.}
}
\description{
Plot customized heatmap.
}
