% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_task_utilities.R
\name{PrepareMatrices}
\alias{PrepareMatrices}
\title{Precompute matrices XTX and XTY for \code{\link{TreeGuidedGroupLasso}}.}
\usage{
PrepareMatrices(Y, X = NULL, task.specific.features = list(),
  standardize = TRUE)
}
\arguments{
\item{Y}{N by K output matrix for every task.}

\item{X}{N by J1 matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 matrix for one particular task
(where columns are features). List has to be ordered according to the
columns of Y.}

\item{standardize}{Standardize data (default is TRUE).}
}
\value{
List containing Y, X, task.specific.features, XTX and XTY. Also
  return means and standard deviations for Y and all inputs.
}
\description{
Computes XTX and XTY. If task specific features are supplied, the result will
be a list containing the respective matrices for each task.
}
\seealso{
\code{\link{TreeGuidedGroupLasso}}, \code{\link{RunGroupCrossvalidation}}.
}
