% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tggl_cv.R
\name{RunGroupCrossvalidation}
\alias{RunGroupCrossvalidation}
\title{Cross-validation wrapper for \code{\link{TreeGuidedGroupLasso}}.}
\usage{
RunGroupCrossvalidation(X = NULL, task.specific.features = list(), Y,
  groups, weights.matrix, lambda.vec, num.folds = 10, num.threads = 1, ...)
}
\arguments{
\item{X}{Column centered N by J input matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{Y}{Column centered N by K output matrix for every task.}

\item{groups}{V by K matrix determining group membership: Task k in group v
iff groups[v,k] == 1 weights.matrix: Numerical matrix where each row is a
numerical vector mapping groups to their weight.}

\item{weights.matrix}{Numerical matrix where each row is a numerical vector
mapping groups to their weight.}

\item{lambda.vec}{Vector of regularization parameters.}

\item{num.folds}{Number of folds.}

\item{num.threads}{Number of threads to use.}

\item{...}{Additional parameters passed to
\code{\link{TreeGuidedGroupLasso}}.}
}
\value{
List containing
    \item{cv.results}{data.frame with cross-validation errors for
    different parameters.}
    \item{full.model}{Full model trained on the whole data set.}
}
\description{
Perform k-fold cross-validation to find optimal parameters settings.
}
\seealso{
\code{\link{TreeGuidedGroupLasso}}
}
