% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_task_by_task.R
\name{RunTBTCrossvalidation}
\alias{RunTBTCrossvalidation}
\title{Cross-validation wrapper for task-by-task lasso model.}
\usage{
RunTBTCrossvalidation(X = NULL, task.specific.features = list(), Y,
  lambda.vec, num.folds = 10, num.threads = 1, ...)
}
\arguments{
\item{X}{N by J input matrix of features common to all tasks.}

\item{task.specific.features}{List of features which are specific to
each task. Each entry contains an N by J2 matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{Y}{N by K output matrix for every task.}

\item{lambda.vec}{Vector of regularization parameters.}

\item{num.folds}{Number of folds.}

\item{num.threads}{Number of threads to use.}

\item{...}{Additional parameters passed to
\code{\link[glmnet]{cv.glmnet}}.}
}
\value{
List containing
    \item{cv.results}{data.frame with cross-validation errors for
    different parameters.}
    \item{full.model}{Full model trained on the whole data set.}
}
\description{
Perform k-fold cross-validation for each task using \code{\link[glmnet]{cv.glmnet}}.
}
