% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth_example.R
\name{TestLinearMTL}
\alias{TestLinearMTL}
\title{Test of tree-guided group lasso on synthetic data.}
\usage{
TestLinearMTL()
}
\description{
Generate synthetic data as in Kim et al. 2010 and apply tree-guided group
lasso and lasso models.
}
