% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TGGLMix_select.R
\name{TGGLMixSelection}
\alias{TGGLMixSelection}
\title{TGGLMixSelection is a convenience function for which calls RunTGGLMix for all
possible combinations of parameters in M.vec and lambda.vec.}
\usage{
TGGLMixSelection(X = NULL, task.specific.features = list(), Y, M.vec,
  lambda.vec, test.ids, groups, weights, verbose = 0, ...)
}
\arguments{
\item{X}{N by J1 matrix of features common to all tasks.}

\item{task.specific.features}{List of features which are specific to each
task. Each entry contains an N by J2 matrix for one particular task (where
columns are features). List has to be ordered according to the columns of
Y.}

\item{Y}{N by K output matrix for every task.}

\item{M.vec}{Vector with number of clusters to be tested.}

\item{lambda.vec}{Vector with lambda values to be tested.}

\item{test.ids}{Mandatory: Indices of data points to be used for testing.}

\item{groups}{Binary V by K matrix determining group membership: Task k in
group v iff groups[v,k] == 1.}

\item{weights}{V dimensional vector with group weights.}

\item{verbose}{(Optional) Integer in {0,1,2}. verbose = 0: No output.
verbose = 1: Print summary at the end of the optimization. verbose = 2:
Print progress during optimization.}

\item{...}{Additional parameters passed to
\code{\link{RunTGGLMix}}.}
}
\value{
List containing
\item{results}{List of TGGLMix models for each parameter setting.}
\item{top.model}{Model with highest predictive likelihood.}
}
\description{
TGGLMixSelection is a convenience function for which calls RunTGGLMix for all
possible combinations of parameters in M.vec and lambda.vec.
}
