% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{EvaluateLinearMTModel}
\alias{EvaluateLinearMTModel}
\title{Evaluate LinearMTL model.}
\usage{
EvaluateLinearMTModel(X = NULL, task.specific.features = list(), Y, B,
  train.idx, test.idx, task.grouping = NULL, feature.names = NULL,
  task.names = NULL, sd.threshold = 1.5, out.dir = NULL)
}
\arguments{
\item{X}{Column centered N by J input matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{Y}{Column centered N by K output matrix for every task.}

\item{B}{J by K matrix of regression coefficients, J = J1 + J2.}

\item{train.idx}{Indices for the training set.}

\item{test.idx}{Indices for the test set.}

\item{task.grouping}{String vector of length K with group names for each
task.}

\item{feature.names}{Feature names.}

\item{task.names}{Task names.}

\item{sd.threshold}{All features with error changes above sd.threshold times
the standard deviation will be taken as important.}

\item{out.dir}{Output directory for plots.}
}
\value{
List containing
    \item{err.out}{Matrix containing training and test correlation and MSE.}
     \item{error.change}{Error changes for excluding features from the
     predictions.}
     \item{regulators.and.targets}{If task.grouping is not NULL: Matrix
     of strings containing top regulators and their targets.}
}
\description{
Compute training and test errors for the given indices as well as error
changes for excluding features. If task.grouping is supplied, identify top
regulators per group and plot clustering of the regression coefficients along
with the most important features. Save plots to directory out.dir or plot
to screen if out.dir is NULL.
}
