% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_task_utilities.R
\name{MTComputeMeanCorrelation}
\alias{MTComputeMeanCorrelation}
\title{Compute (mean) correlation for linear multi-task model.}
\usage{
MTComputeMeanCorrelation(Y, B, X = NULL, task.specific.features = list(),
  pred = NULL, method = "spearman")
}
\arguments{
\item{Y}{Column centered N by K output matrix for every task.}

\item{B}{J by K matrix of regression coefficients, J = J1 + J2.}

\item{X}{N by J1 matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{pred}{Predicted output matrix. If NULL, compute predictions using input features.}

\item{method}{Correlation method to use.}
}
\value{
The mean correlation between predictions for each task and Y.
}
\description{
Compute (mean) correlation for linear multi-task model.
}
