% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_task_utilities.R
\name{PrepareMatrices}
\alias{PrepareMatrices}
\title{Precompute matrices XTX and XTY for \code{\link{TreeGuidedGroupLasso}}.}
\usage{
PrepareMatrices(Y, X = NULL, task.specific.features = list(), idx = NULL)
}
\arguments{
\item{Y}{Column centered N by K output matrix for every task.}

\item{X}{Column-centered N by J1 matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{idx}{Vector of indices (data points) to exclude.}
}
\value{
List with XTX and XTY if no task specific features are supplied, or
  list of lists otherwise.
}
\description{
Computes XTX and XTY. If task specific features are supplied, the result will
be a list containing the respective matrices for each task.
}
\seealso{
\code{\link{TreeGuidedGroupLasso}}, \code{\link{RunGroupCrossvalidation}}.
}
