% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_task_by_task.R
\name{RunTBTCrossvalidation}
\alias{RunTBTCrossvalidation}
\title{Cross-validation wrapper for task-by-task lasso model.}
\usage{
RunTBTCrossvalidation(X = NULL, task.specific.features = list(), Y,
  lambda.vec, num.folds = 10, num.threads = 1, ...)
}
\arguments{
\item{X}{Column centered N by J input matrix of features common to all tasks.}

\item{task.specific.features}{Named list of features which are specific to
each task. Each entry contains an N by J2 column-centered matrix for one
particular task (where columns are features). List has to be ordered
according to the columns of Y.}

\item{Y}{Column centered N by K output matrix for every task.}

\item{lambda.vec}{Vector of regularization parameters.}

\item{num.folds}{Number of folds.}

\item{num.threads}{Number of threads to use.}

\item{...}{Additional parameters passed to
\code{\link[glmnet]{cv.glmnet}}.}
}
\value{
List containing
   \item{lambda}{Best lambda for each task.}
   \item{beta}{Final regression coefficients for the model fitted on the full data set.}
   \item{error}{Cross-validation errors for each task.}
}
\description{
Perform k-fold cross-validation for each task using \code{\link[glmnet]{cv.glmnet}}.
}
